Refactor Entity & EntityData classes
