+++ Celbridge/Modules/Workspace/Celbridge.Entities/ServiceConfiguration.cs
@@ -1,6 +1,5 @@
 using Celbridge.Modules;
 using Celbridge.Entities.Services;
-using Celbridge.Entities.Models;
 
 namespace Celbridge.Entities;
 
@@ -14,10 +13,5 @@ public static class ServiceConfiguration
         config.AddTransient<IEntityService, EntityService>();
         config.AddTransient<EntitySchemaService>();
         config.AddTransient<EntityPrototypeService>();
-
-        //
-        // Register models
-        //
-        config.AddTransient<Entity>();
     }
 }
